Added a bunch of debug messages in the hopes of debugging. Currently crashes when trying to deal with the calibration: Line 83 EventLoaderALiBaVa.cpp Line 201 auxfunctions.cpp Line 119 auxfunctions.cpp
Probably a badly made TH1 object:
Line 199 auxfunctions.cpp
Line 74 auxfunctions.cpp

_____________
IGNORE THE ABOVE:
It crashes because it tries to access a TH1 object which doesn't exist.
The TH1 object is only created if the calibration file is of type (which it should be).
The file is of type 0 for some reason.